+++ lib/widgets/screens/search_results_screen.dart
@@ -1,31 +1,58 @@
 import 'package:flutter/material.dart';
 
 import '../../models/search.dart';
+import '../../services/observer_utils.dart';
 import 'search_screen.dart';
 
 class SearchResultsScreen extends StatefulWidget {
   final Search search;
   final String? title;
   final bool showSummary;
+  final bool autoUpdate;
   const SearchResultsScreen({
     super.key,
     required this.search,
     this.title,
     this.showSummary = true,
+    this.autoUpdate = false,
   });
 
   @override
   State<SearchResultsScreen> createState() => _SearchResultsScreenState();
 }
 
-class _SearchResultsScreenState extends State<SearchResultsScreen> {
+class _SearchResultsScreenState extends State<SearchResultsScreen>
+    with RouteAware {
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    if (widget.autoUpdate) {
+      ObserverUtils.routeObserver.subscribe(this, ModalRoute.of(context)!);
+    }
+  }
+
+  @override
+  void dispose() {
+    if (widget.autoUpdate) {
+      ObserverUtils.routeObserver.unsubscribe(this);
+    }
+    super.dispose();
+  }
+
+  @override
+  void didPopNext() async {
+    return await _refreshScreen();
+  }
+
+  Future<void> _refreshScreen() async {
+    await widget.search.resetChildren();
+    if (context.mounted) setState(() {});
+  }
+
   @override
   Widget build(BuildContext context) {
     return RefreshIndicator(
-      onRefresh: () async {
-        await widget.search.resetChildren();
-        if (context.mounted) setState(() {});
-      },
+      onRefresh: _refreshScreen,
       child: CustomScrollView(
         // cacheExtent: 400.0,
         slivers: <Widget>[