+++ lib/widgets/screens/updates_screen.dart
@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 
 import '../../models/search_parameters.dart';
 import '../../models/updates.dart';
+import '../../services/observer_utils.dart';
 import 'search_screen.dart';
 
 class UpdatesScreen extends StatefulWidget {
@@ -15,14 +16,33 @@ class UpdatesScreen extends StatefulWidget {
   State<UpdatesScreen> createState() => _UpdatesScreenState();
 }
 
-class _UpdatesScreenState extends State<UpdatesScreen> {
+class _UpdatesScreenState extends State<UpdatesScreen> with RouteAware {
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    ObserverUtils.routeObserver.subscribe(this, ModalRoute.of(context)!);
+  }
+
+  @override
+  void dispose() {
+    ObserverUtils.routeObserver.unsubscribe(this);
+    super.dispose();
+  }
+
+  @override
+  void didPopNext() async {
+    await _refreshScreen();
+  }
+
+  Future<void> _refreshScreen() async {
+    await widget.updates.resetChildren();
+    if (context.mounted) setState(() {});
+  }
+
   @override
   Widget build(BuildContext context) {
     return RefreshIndicator(
-      onRefresh: () async {
-        await widget.updates.resetChildren();
-        if (context.mounted) setState(() {});
-      },
+      onRefresh: _refreshScreen,
       child: CustomScrollView(
         // cacheExtent: 400.0,
         slivers: <Widget>[